requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg[binary]==3.2.3
django-cors-headers==4.3.1
Pillow==10.4.0
